+++ crates/datadog-agent-config/src/deserializers/mod.rs
@@ -1,6 +1,7 @@
 pub mod additional_endpoints;
 pub mod apm_replace_rule;
 pub mod flush_strategy;
+pub mod helpers;
 pub mod log_level;
 pub mod logs_additional_endpoints;
 pub mod processing_rule;
